feat: added a simple board view if the board exists in the database
It is now possible to view a board if the boardId is present in the database (mongodb). It will fetch all data available of that board and show it in a (not final) overview.

The application crashes if you fill in a not valid id in the url. will be fixed when user authentication gets implemented